backend/requirements.txt
Flask>=3.1,<3.2
Flask-Bcrypt>=1.0,<1.1
Flask-Cors>=5.0,<5.1
Flask-JWT-Extended>=4.7,<4.8
Flask-Migrate>=4.0,<4.1
Flask-SQLAlchemy>=3.1,<3.2
Flask-SocketIO>=5.4,<5.5
marshmallow>=3.22,<3.23
PyMySQL>=1.1,<1.2
python-dotenv>=1.0,<1.1
pytest>=8.3,<8.4
